Move "sections" folder to under the "components" folder
I prefer to keep all my React components under the same folder.  Previously there were two folders next to each other named "sections" and "components" which both contained React components.  This resolves that confusion by nesting the "sections" folder under the "components" folder.
